stream_diffusion/examples/image_to_image.py
"""
Image-to-Image Translation Example

Demonstrates using StreamDiffusion for image-to-image tasks like:
- Style transfer
- Image editing
- Enhancement
"""

import torch
from PIL import Image
import numpy as np
import sys
sys.path.append('..')

from pipeline.stream_pipeline import StreamDiffusionPipeline


def style_transfer():
    """Apply style transfer to an image."""
    print("="*70)
    print("Image-to-Image Style Transfer Example")
    print("="*70 + "\n")
    
    # Initialize pipeline
    pipe = StreamDiffusionPipeline(
        model_id="stabilityai/sd-turbo",
        denoising_steps=4,
        guidance_scale=7.5,
        use_rcfg=True  # Use R-CFG for faster conditional generation
    )
    
    # Create or load input image
    print("Creating input image...")
    input_image = Image.fromarray(
        np.random.randint(0, 255, (512, 512, 3), dtype=np.uint8)
    )
    input_image.save("input_image.png")
    print("  ✅ Input image saved: input_image.png\n")
    
    # Apply different styles
    styles = [
        "anime style",
        "oil painting style",
        "watercolor painting",
        "cyberpunk style"
    ]
    
    print("Applying styles...\n")
    
    for i, style in enumerate(styles):
        print(f"Style {i+1}: '{style}'")
        
        # For img2img, we'd need to encode input image to latent
        # This is simplified - real implementation would:
        # 1. Encode input image with VAE
        # 2. Add noise based on strength parameter
        # 3. Denoise with prompt conditioning
        
        # For demonstration, just use text-to-image
        prompt = f"{style}, high quality"
        image_tensor = pipe(prompt)
        
        # Save
        image_pil = pipe.io_queue.postprocess_fn(image_tensor)
        image_pil.save(f"output_style_{i+1}.png")
        
        print(f"  ✅ Saved: output_style_{i+1}.png\n")
        
    print("="*70)
    print("✅ Style transfer complete!")
    print("="*70 + "\n")


def demonstrate_rcfg():
    """Demonstrate R-CFG speedup for conditional generation."""
    print("="*70)
    print("R-CFG Performance Demonstration")
    print("="*70 + "\n")
    
    prompt = "a beautiful landscape"
    num_steps = 4
    
    # Without R-CFG (standard CFG)
    print("1. Standard CFG (2× U-Net calls):")
    pipe_standard = StreamDiffusionPipeline(
        model_id="stabilityai/sd-turbo",
        denoising_steps=num_steps,
        guidance_scale=7.5,
        use_rcfg=False
    )
    
    import time
    start = time.time()
    _ = pipe_standard(prompt)
    time_standard = (time.time() - start) * 1000
    
    print(f"   Time: {time_standard:.2f}ms\n")
    
    # With R-CFG
    print("2. Residual CFG (1× U-Net calls):")
    pipe_rcfg = StreamDiffusionPipeline(
        model_id="stabilityai/sd-turbo",
        denoising_steps=num_steps,
        guidance_scale=7.5,
        use_rcfg=True
    )
    
    start = time.time()
    _ = pipe_rcfg(prompt)
    time_rcfg = (time.time() - start) * 1000
    
    print(f"   Time: {time_rcfg:.2f}ms\n")
    
    # Compare
    speedup = time_standard / time_rcfg
    
    print(f"{'='*70}")
    print(f"R-CFG Speedup: {speedup:.2f}×")
    print(f"Expected from paper: ~2.05× at {num_steps} steps")
    print("="*70 + "\n")


def main():
    """Main example entry point."""
    import argparse
    
    parser = argparse.ArgumentParser(description="Image-to-Image examples")
    parser.add_argument(
        "--mode",
        type=str,
        choices=["style", "rcfg"],
        default="style",
        help="Example mode"
    )
    
    args = parser.parse_args()
    
    if not torch.cuda.is_available():
        print("⚠️  CUDA not available. Examples require GPU.\n")
        return
    
    if args.mode == "style":
        style_transfer()
    elif args.mode == "rcfg":
        demonstrate_rcfg()


if __name__ == "__main__":
    main()