stream_diffusion/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.7.0
tqdm>=4.65.0
matplotlib>=3.7.0
scipy>=1.10.0

# Optional for maximum performance
# tensorrt>=8.6.0
# onnx>=1.14.0